+++ app/Actions/CategoryActions/ShowAction.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Actions\CategoryActions;
+
+use App\Models\Category;
+
+class ShowAction
+{
+    public static function execute(int $id): Category
+    {
+        return Category::findOrFail($id);
+    }
+}