+++ app/Actions/LinkActions/UpdateAction.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace App\Actions\LinkActions;
+
+use App\Models\Link;
+
+class UpdateAction
+{
+    public static function execute(Link $link, array $data): Link
+    {
+        $link->update([
+            ...$data,
+            'url_hash' => isset($data['url']) ? hash('sha256', $data['url']) : $link->url_hash,
+        ]);
+
+        return $link->refresh();
+    }
+}