+++ app/Actions/TagActions/ShowAction.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Actions\TagActions;
+
+use App\Models\Tag;
+
+class ShowAction
+{
+    public static function execute(int $id): Tag
+    {
+        return Tag::findOrFail($id);
+    }
+}