+++ app/Filament/Resources/Tags/TagResource.php
@@ -11,9 +11,9 @@ use App\Filament\Resources\Tags\Schemas\TagInfolist;
 use App\Filament\Resources\Tags\Tables\TagsTable;
 use App\Models\Tag;
 use BackedEnum;
+use Daljo25\FilamentTablerIcons\Enums\TablerIcon;
 use Filament\Resources\Resource;
 use Filament\Schemas\Schema;
-use Daljo25\FilamentTablerIcons\Enums\TablerIcon;
 use Filament\Tables\Table;
 
 class TagResource extends Resource